Application/requirements.txt
beautifulsoup4==4.12.3
langchain==0.1.17
langchain_community==0.0.36
langchain_core==0.1.50
langchain_openai==0.1.6
langchain_text_splitters==0.0.1
numpy==1.26.4
pandas==2.2.2
python_pptx==0.6.23
Requests==2.31.0
streamlit==1.12.0
